requirements.txt
google-cloud-dialogflow==2.14.1
python-dotenv==0.20.0
python-telegram-bot==13.2
vk-api==11.9.8
